Lab1_6 completed

+++ Lab1/Lab1_6/Lab1_6.cpp
@@ -7,8 +7,8 @@ int main()
 	const int sec_In_Min = 60;
 	int sec_In_Hour = 60 * sec_In_Min;
 	int sec_In_Day = 24 * sec_In_Hour;
-	int sec_In_Year = 365 * sec_In_Day;
-	long long sec_In_100_Years = 100 * sec_In_Year; // переполнение
-	long long sec_In_1000_Years = 1000 * sec_In_Year; // переполнение
+	long long sec_In_Year = 365 * sec_In_Day;
+	long long sec_In_100_Years = 100 * sec_In_Year; 
+	long long sec_In_1000_Years = 1000 * sec_In_Year; 
 }
